refactor: moved `bonusVacanze/components/markdown`
