+++ ts/components/markdown/FooterSingleButton.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
-import FooterWithButtons from "../../../../../components/ui/FooterWithButtons";
-import { cancelButtonProps } from "../../../../../components/buttons/ButtonConfigurations";
+import FooterWithButtons from "../ui/FooterWithButtons";
+import { cancelButtonProps } from "../buttons/ButtonConfigurations";
 
 export type OnCancelProps = {
   onCancel: () => void;
