+++ ts/features/bonus/bpd/screens/iban/ko/IbanKONotOwned.tsx
@@ -14,7 +14,7 @@ import BaseScreenComponent from "../../../../../../components/screens/BaseScreen
 import I18n from "../../../../../../i18n";
 import { profileSelector } from "../../../../../../store/reducers/profile";
 import { GlobalState } from "../../../../../../store/reducers/types";
-import { FooterTwoButtons } from "../../../../bonusVacanze/components/markdown/FooterTwoButtons";
+import { FooterTwoButtons } from "../../../../../../components/markdown/FooterTwoButtons";
 import {
   bpdIbanInsertionCancel,
   bpdIbanInsertionResetScreen
