+++ ts/features/wallet/onboarding/bancomatPay/screens/searchBPay/BPayKoTimeout.tsx
@@ -9,7 +9,7 @@ import { InfoScreenComponent } from "../../../../../../components/infoScreen/Inf
 import BaseScreenComponent from "../../../../../../components/screens/BaseScreenComponent";
 import I18n from "../../../../../../i18n";
 import { GlobalState } from "../../../../../../store/reducers/types";
-import { FooterTwoButtons } from "../../../../../bonus/bonusVacanze/components/markdown/FooterTwoButtons";
+import { FooterTwoButtons } from "../../../../../../components/markdown/FooterTwoButtons";
 import { searchUserBPay, walletAddBPayCancel } from "../../store/actions";
 import { onboardingBPayAbiSelectedSelector } from "../../store/reducers/abiSelected";
