+++ ts/features/wallet/onboarding/cobadge/screens/search/ko/CoBadgeKoTimeout.tsx
@@ -11,7 +11,7 @@ import I18n from "../../../../../../../i18n";
 import { GlobalState } from "../../../../../../../store/reducers/types";
 import { emptyContextualHelp } from "../../../../../../../utils/emptyContextualHelp";
 import { useHardwareBackButton } from "../../../../../../../hooks/useHardwareBackButton";
-import { FooterTwoButtons } from "../../../../../../bonus/bonusVacanze/components/markdown/FooterTwoButtons";
+import { FooterTwoButtons } from "../../../../../../../components/markdown/FooterTwoButtons";
 import {
   searchUserCoBadge,
   walletAddCoBadgeCancel
